refactor: rename `SignUpReducer` to SignUpLogic
